Set up initial M2 scene, lengthened stem mine anim
