+++ Assets/Scripts/Graphics and UI/Interaction/TutorialButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6573ea5f34724d72ace2569cdc75f9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
