+++ Assets/Scripts/Graphics and UI/Interaction/TutorialDoor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 495401f9b7b3f41e9b022e3b59308979
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
