--- Assets/TutorialButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f6573ea5f34724d72ace2569cdc75f9e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
