+++ UserSettings/EditorUserSettings.asset
@@ -6,28 +6,28 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 22424703114646680e0b0227036c7e1c130f57072524232339261336a2b27a2decee22f0
+      value: 22424703114646680e0b0227036c7d0217191c2522670b626d0b0f32ece43b36acf238e0f323
       flags: 0
     RecentlyUsedScenePath-1:
-      value: 224247031146466f181a03301922530359301d24293a273c621a1e36ece57a2decee22f0
+      value: 22424703114646680e0b0227036c7c18041e0b65382d3524633c133af6f9
       flags: 0
     RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c7a1c1a1e173e3867023520265d62acf53a31f6fe
       flags: 0
     RecentlyUsedScenePath-3:
-      value: 22424703114646680e0b0227036c7c18041e0b65382d3524633c133af6f9
-      flags: 0
-    RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c7d0217191c252267023520265d62acf53a31f6fe
-      flags: 0
-    RecentlyUsedScenePath-5:
-      value: 22424703114646680e0b0227036c7d0217191c25226704222c27193cecae2136ebf32f
-      flags: 0
-    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c72191a120b3e232623707c67083debf42d
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c72191a120b3e232623707f67083debf42d
+      flags: 0
+    RecentlyUsedScenePath-5:
+      value: 22424703114646680e0b0227036c7133171e570e292529700120093aedee7a2decee22f0
+      flags: 0
+    RecentlyUsedScenePath-6:
+      value: 22424703114646680e0b0227036c6f190214106a082d2b3f633c133af6f9
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 22424703114646680e0b0227036c6b15050311242b0f343f38271920add22d39eca923e7ee2e26
       flags: 0
     RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c7119151c0165183d323f3f201c3facf53a31f6fe
